+++ src/GafferScene/BranchCreator.cpp
@@ -108,6 +108,24 @@ bool internedStringValueLess( const InternedString &a, const InternedString &b )
 	return a.string() < b.string();
 }
 
+const InternedString g_ellipsis( "..." );
+
+void validateDestination( const ScenePlug::ScenePath &destination )
+{
+	for( auto &n : destination )
+	{
+		if( n == g_ellipsis )
+		{
+			throw IECore::Exception(
+				fmt::format(
+					"Invalid destination `{}` (must not contain `...`)",
+					ScenePlug::pathToString( destination )
+				)
+			);
+		}
+	}
+}
+
 } // namespace
 
 //////////////////////////////////////////////////////////////////////////
@@ -301,7 +319,9 @@ class BranchCreator::BranchesData : public IECore::Data
 			h.append( path.data(), path.size() );
 			h.append( (uint64_t)path.size() );
 
-			ScenePlug::ScenePath destination = ScenePlug::stringToPath( branchCreator->destinationPlug()->getValue() );
+			const ScenePlug::ScenePath destination = ScenePlug::stringToPath( branchCreator->destinationPlug()->getValue() );
+			validateDestination( destination );
+
 			h.append( destination.data(), destination.size() );
 			h.append( (uint64_t)destination.size() );
 
@@ -312,7 +332,8 @@ class BranchCreator::BranchesData : public IECore::Data
 
 		void addBranch( const BranchCreator *branchCreator, const ScenePlug::ScenePath &path )
 		{
-			ScenePlug::ScenePath destination = ScenePlug::stringToPath( branchCreator->destinationPlug()->getValue() );
+			const ScenePlug::ScenePath destination = ScenePlug::stringToPath( branchCreator->destinationPlug()->getValue() );
+			validateDestination( destination );
 			const ScenePlug::ScenePath existing = closestExistingPath( branchCreator->inPlug(), destination );
 
 			tbb::spin_mutex::scoped_lock lock( m_mutex );